+++ plugins/CoreHome/javascripts/dataTable_rowactions.js
@@ -386,6 +386,11 @@ DataTable_RowActions_RowEvolution.prototype.showRowEvolution = function (apiMeth
     requestParams.action = 'getRowEvolutionPopover';
     requestParams.colors = JSON.stringify(piwik.getSparklineColors());
 
+    var idDimension = broadcast.getValueFromHash('idDimension');
+    if (idDimension) {
+        requestParams.idDimension = parseInt(idDimension, 10);
+    }
+
     $.extend(requestParams, extraParams);
 
     var ajaxRequest = new ajaxHelper();